organize different message validations in their own files
